requirements.txt
datasets==2.21.0
numpy==2.0.1
pandas==2.2.2
Pillow==10.4.0
scikit_learn==1.5.1
torch==2.4.0
torchvision==0.19.0
